group logic fix
previously when an incorrect group close token occures (for example: '( te nhron ]'),
an error is pushed, the group is closed as if the close token was correct, and the close token is consumed.

this is bad because if we have: "{ [ }" we only want to show 1 error which is that the [ was not closed, but under the previous rules the } is consumed and an unexpected end of file error is also pushed.

this commit replaces the logic so that incorrect group close tokens are not consumed and try to close another group, which eliminates the second unnecessary error.