format system
finished.

a format is an output form of a shader (wgsl, spirv...) which can be evaluated from a pipeline. you can call the format fn to get the pipeline in any format, and you can call the const optimize fn for a specific format so that it will do as much of the work possible at cp. so even right now in stable rust not alot is possible, in the future where we will have more const features translating into formats can become much faster without any change from the user of the crate.